requirements.txt
Appium-Python-Client
selenium
pytest
pytest-env
pytest-rerunfailures
allure-pytest
